Started reworking the landing page for a better look
