+++ package.json
@@ -7,14 +7,15 @@
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
     "@internationalized/date": "^3.8.0",
-    "@mui/icons-material": "^6.4.8",
-    "@mui/material": "^6.4.8",
+    "@mui/icons-material": "^7.1.1",
+    "@mui/material": "^7.1.1",
     "@mui/x-date-pickers": "^7.28.3",
     "@react-aria/example-theme": "^1.0.5",
     "@testing-library/dom": "^10.4.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.2.0",
     "@testing-library/user-event": "^13.5.0",
+    "@toolpad/core": "^0.15.0",
     "@turf/turf": "^7.2.0",
     "date-fns": "^4.1.0",
     "dayjs": "^1.11.13",
