+++ src/components/LandingPageHeader.js
@@ -1,0 +1,9 @@
+export default function LandingPageHeader() {
+    return (
+        <div className="landing-page-header">
+            Welcome to 4DGeo!
+            <p>Explore our example dashboards or create your own custom dashboard.</p>
+            <p>Click on a template to get started.</p>
+        </div>
+    )
+}
