+++ src/index.js
@@ -9,6 +9,7 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
+    <link rel="icon" href="./4dgeo/3dgeo.ico" />
     <BrowserRouter basename='/4dgeo'>
       <Routes>
         <Route path="/" element={<LandingPage />} />
